separate code into EnvoyServerProtoData as per review comment
