+++ xds/src/main/java/io/grpc/xds/XdsClient.java
@@ -32,9 +32,9 @@ import io.grpc.internal.ObjectPool;
 import io.grpc.xds.Bootstrapper.ChannelCreds;
 import io.grpc.xds.Bootstrapper.ServerInfo;
 import io.grpc.xds.EnvoyProtoData.DropOverload;
-import io.grpc.xds.EnvoyProtoData.Listener;
 import io.grpc.xds.EnvoyProtoData.Locality;
 import io.grpc.xds.EnvoyProtoData.LocalityLbEndpoints;
+import io.grpc.xds.EnvoyServerProtoData.Listener;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashMap;
